WIP  - feature #354 Update Integration to handle files from API   - feature #354 Update Integration to handle files from API
