+++ gendox-core-api/pom.xml
@@ -408,6 +408,10 @@
             <artifactId>jtokkit</artifactId>
             <version>1.1.0</version>
         </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-annotations</artifactId>
+        </dependency>
     </dependencies>
 
     <dependencyManagement>
