+++ gendox-core-api/src/main/java/dev/ctrlspace/gendox/gendoxcoreapi/repositories/ProjectDocumentRepository.java
@@ -35,10 +35,9 @@ public interface ProjectDocumentRepository extends JpaRepository<ProjectDocument
 
     Optional<ProjectDocument> findByDocumentIdAndProjectId(UUID documentId, UUID projectId);
 
-    // Custom query to retrieve a projectId based on documentId
-    @Query("SELECT pd.project.id FROM ProjectDocument pd WHERE pd.documentId = :documentId LIMIT 1")
-    Optional<UUID> findProjectIdByDocumentId(UUID documentId);
 
+    @Query(value = "SELECT pd.project_id FROM project_document pd WHERE pd.document_id = :documentId LIMIT 1", nativeQuery = true)
+    Optional<UUID> findProjectIdByDocumentId(UUID documentId);
 
 
 }
