+++ gendox-core-api/src/main/java/dev/ctrlspace/gendox/spring/batch/jobs/splitter/steps/DocumentSplitterProcessor.java
@@ -1,9 +1,9 @@
 package dev.ctrlspace.gendox.spring.batch.jobs.splitter.steps;
 
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import dev.ctrlspace.gendox.gendoxcoreapi.exceptions.GendoxException;
 import dev.ctrlspace.gendox.gendoxcoreapi.model.DocumentInstance;
 import dev.ctrlspace.gendox.gendoxcoreapi.model.ProjectAgent;
-import dev.ctrlspace.gendox.gendoxcoreapi.services.DocumentService;
 import dev.ctrlspace.gendox.gendoxcoreapi.services.DownloadService;
 import dev.ctrlspace.gendox.gendoxcoreapi.services.ProjectAgentService;
 import dev.ctrlspace.gendox.gendoxcoreapi.utils.CryptographyUtils;
@@ -28,6 +28,9 @@ import java.util.List;
 public class DocumentSplitterProcessor implements ItemProcessor<DocumentInstance, DocumentSectionDTO> {
 
     Logger logger = LoggerFactory.getLogger(DocumentSplitterProcessor.class);
+    @Value("#{jobParameters['skipUnchangedDocs']}")
+    protected Boolean skipUnchangedDocs;
+
     private ServiceSelector serviceSelector;
     private ProjectAgentService projectAgentService;
     private DownloadService downloadService;
@@ -35,21 +38,17 @@ public class DocumentSplitterProcessor implements ItemProcessor<DocumentInstance
     private CryptographyUtils cryptographyUtils;
 
 
-    @Value("#{jobParameters['skipUnchangedDocs']}")
-    protected Boolean skipUnchangedDocs;
-
     @Autowired
     public DocumentSplitterProcessor(ServiceSelector serviceSelector,
                                      ProjectAgentService projectAgentService,
                                      DownloadService downloadService,
+                                     GendoxAPIIntegrationService gendoxAPIIntegrationService,
                                      CryptographyUtils cryptographyUtils) {
-                                     DownloadService downloadService,
-                                     GendoxAPIIntegrationService gendoxAPIIntegrationService) {
         this.serviceSelector = serviceSelector;
         this.projectAgentService = projectAgentService;
         this.downloadService = downloadService;
-        this.cryptographyUtils = cryptographyUtils;
         this.gendoxAPIIntegrationService = gendoxAPIIntegrationService;
+        this.cryptographyUtils = cryptographyUtils;
     }
 
     String baseUrl = "https://test.dma.com.gr/wp-json";
@@ -64,7 +63,7 @@ public class DocumentSplitterProcessor implements ItemProcessor<DocumentInstance
 
 
         try {
-            String fileContent = null;
+            String fileContent = downloadService.readDocumentContent(item.getRemoteUrl());
 
             fileContent = (item.getFileType() == null || !"API_INTEGRATION_FILE".equals(item.getFileType().getName()))
                     ? downloadService.readDocumentContent(item.getRemoteUrl())
@@ -73,9 +72,6 @@ public class DocumentSplitterProcessor implements ItemProcessor<DocumentInstance
             if (item.getFileType() == null) {
                 logger.warn("DocumentInstance {} has a null fileType. Using remoteUrl to retrieve content.", item.getId());
             }
-
-
-            String fileContent = downloadService.readDocumentContent(item.getRemoteUrl());
 
             // SHA-256 hash check only if splitAllDocuments is false
             if (Boolean.TRUE.equals(skipUnchangedDocs)) {
@@ -88,32 +84,31 @@ public class DocumentSplitterProcessor implements ItemProcessor<DocumentInstance
                     return null;
                 }
 
-                // Update hash to the new value since content has changed
                 item.setDocumentSha256Hash(documentSha256Hash);
                 documentUpdated = true;
             } else {
                 logger.trace("splitAllDocuments is true, skipping SHA-256 check for document {}.", item.getId());
             }
 
-            // Fetch the project agent by document ID
+
+
+
             agent = projectAgentService.getAgentByDocumentId(item.getId());
             String splitterTypeName = agent.getDocumentSplitterType().getName();
-
-            // Get the document splitter by name
             DocumentSplitter documentSplitter = serviceSelector.getDocumentSplitterByName(splitterTypeName);
 
             if (documentSplitter == null) {
-                throw new GendoxException("DOCUMENT_SPLITTER_NOT_FOUND",
-                        "Document splitter not found with name: " + splitterTypeName, HttpStatus.NOT_FOUND);
+                throw new GendoxException("DOCUMENT_SPLITTER_NOT_FOUND", "Document splitter not found with name: " + splitterTypeName, HttpStatus.NOT_FOUND);
             }
 
             contentSections = documentSplitter.split(fileContent);
 
-
-        } catch (Exception e) {
-            logger.warn("Error {} splitting document to sections {}. Skipping...", e.getMessage(), item.getId());
+        } catch (
+                Exception e) {
+            logger.warn("Error {} split document to sections {}. Skipping...", e.getMessage(), item.getId());
             return null;
         }
+
 
         return new DocumentSectionDTO(item, contentSections, documentUpdated);
     }
